requirements.txt
pandas>=2.2,<3
numpy>=1.26,<3
scikit-learn>=1.4,<2
streamlit>=1.33,<2
plotly>=5.20,<7
joblib>=1.4,<2
tqdm>=4.66,<5
lightgbm>=4.3,<5
xgboost>=2.0,<3
ta>=0.11,<1
datasets>=4.5,<5
google-genai>=1.63,<2
python-dotenv>=1.2,<2
yfinance>=0.2,<1
fpdf2>=2.8.6,<3.0.0
alpaca-py>=0.38,<1
requests>=2.31,<3
